shadowManager: fix deadlock in basicShadowSubscribe
If the MQTT connection is temporarily lost, subscribe will throw an
exception and leave _shadowSubUnsubOperationLock locked.  The next
attempt to lock this lock will deadlock, preventing any future use
of the shadow.

Release the lock when an exception occurs so that the shadow can
continue to be used after the MQTT connection auto-reconnects.